+++ frontend/.size-limit.js
@@ -22,6 +22,6 @@ module.exports = [
     name: 'First Load JS (total)',
     path: '.next/static/chunks/**/*.js',
     gzip: true,
-    limit: '1.75 MB',
+    limit: '1.76 MB',
   },
 ];